% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_gz_file}
\alias{read_gz_file}
\title{Read contents of .gz file(s) into a single \code{character()}.}
\usage{
read_gz_file(file_paths)
}
\arguments{
\item{file_paths}{File path from which to read.}
}
\value{
\code{character()}
}
\description{
Read contents of .gz file(s) into a single \code{character()}.
}
\details{
If more than one file path is provided, each elemente of the returned vector contains the contents of a single file.
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
