% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_gz_lines}
\alias{read_gz_lines}
\title{Read contents of .gz file(s) into a \code{character()}.}
\usage{
read_gz_lines(file_paths)
}
\arguments{
\item{file_paths}{File path from which to read.}
}
\value{
\code{character()}
}
\description{
Read contents of .gz file(s) into a \code{character()}.
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
